Create README.md
To do List project - developed from the Formação Full Stack Javascript Thiago M. Medeiros' course to present the new skills I got in JavaScript

features of the project:
Scheduled with Javascript only, with no frameworks. 
Introduce an activity, mark it as 'ok' and delete it.
keyboard event to load up the work to the list.